// web/src/api/models/TypesKernelRoute.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesKernelRoute
 */
export interface TypesKernelRoute {
    /**
     * 
     * @type {string}
     * @memberof TypesKernelRoute
     */
    dev?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesKernelRoute
     */
    dst: string;
    /**
     * 
     * @type {string}
     * @memberof TypesKernelRoute
     */
    family: string;
    /**
     * 
     * @type {string}
     * @memberof TypesKernelRoute
     */
    gateway?: string;
    /**
     * 
     * @type {number}
     * @memberof TypesKernelRoute
     */
    metric?: number;
    /**
     * 
     * @type {string}
     * @memberof TypesKernelRoute
     */
    protocol: string;
}

/**
 * Check if a given object implements the TypesKernelRoute interface.
 */
export function instanceOfTypesKernelRoute(value: object): value is TypesKernelRoute {
    if (!('dst' in value) || value['dst'] === undefined) return false;
    if (!('family' in value) || value['family'] === undefined) return false;
    if (!('protocol' in value) || value['protocol'] === undefined) return false;
    return true;
}

export function TypesKernelRouteFromJSON(json: any): TypesKernelRoute {
    return TypesKernelRouteFromJSONTyped(json, false);
}

export function TypesKernelRouteFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesKernelRoute {
    if (json == null) {
        return json;
    }
    return {
        
        'dev': json['dev'] == null ? undefined : json['dev'],
        'dst': json['dst'],
        'family': json['family'],
        'gateway': json['gateway'] == null ? undefined : json['gateway'],
        'metric': json['metric'] == null ? undefined : json['metric'],
        'protocol': json['protocol'],
    };
}

export function TypesKernelRouteToJSON(json: any): TypesKernelRoute {
    return TypesKernelRouteToJSONTyped(json, false);
}

export function TypesKernelRouteToJSONTyped(value?: TypesKernelRoute | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'dev': value['dev'],
        'dst': value['dst'],
        'family': value['family'],
        'gateway': value['gateway'],
        'metric': value['metric'],
        'protocol': value['protocol'],
    };
}