// web/src/api/models/TypesOSPFNeighborSummary.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Routier API
 * Configuration and monitoring API for a Routier node.
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TypesOSPFNeighborSummary
 */
export interface TypesOSPFNeighborSummary {
    /**
     * 
     * @type {string}
     * @memberof TypesOSPFNeighborSummary
     */
    address: string;
    /**
     * 
     * @type {string}
     * @memberof TypesOSPFNeighborSummary
     */
    area?: string;
    /**
     * 
     * @type {string}
     * @memberof TypesOSPFNeighborSummary
     */
    interface: string;
    /**
     * 
     * @type {string}
     * @memberof TypesOSPFNeighborSummary
     */
    neighbor_id: string;
    /**
     * 
     * @type {string}
     * @memberof TypesOSPFNeighborSummary
     */
    state: string;
}

/**
 * Check if a given object implements the TypesOSPFNeighborSummary interface.
 */
export function instanceOfTypesOSPFNeighborSummary(value: object): value is TypesOSPFNeighborSummary {
    if (!('address' in value) || value['address'] === undefined) return false;
    if (!('interface' in value) || value['interface'] === undefined) return false;
    if (!('neighbor_id' in value) || value['neighbor_id'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    return true;
}

export function TypesOSPFNeighborSummaryFromJSON(json: any): TypesOSPFNeighborSummary {
    return TypesOSPFNeighborSummaryFromJSONTyped(json, false);
}

export function TypesOSPFNeighborSummaryFromJSONTyped(json: any, ignoreDiscriminator: boolean): TypesOSPFNeighborSummary {
    if (json == null) {
        return json;
    }
    return {
        
        'address': json['address'],
        'area': json['area'] == null ? undefined : json['area'],
        'interface': json['interface'],
        'neighbor_id': json['neighbor_id'],
        'state': json['state'],
    };
}

export function TypesOSPFNeighborSummaryToJSON(json: any): TypesOSPFNeighborSummary {
    return TypesOSPFNeighborSummaryToJSONTyped(json, false);
}

export function TypesOSPFNeighborSummaryToJSONTyped(value?: TypesOSPFNeighborSummary | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'address': value['address'],
        'area': value['area'],
        'interface': value['interface'],
        'neighbor_id': value['neighbor_id'],
        'state': value['state'],
    };
}